% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{classified_landscape}
\alias{classified_landscape}
\title{Example map (factor).}
\format{A raster layer object.}
\source{
Simulated neutral landscape models with R. \url{https://github.com/ropensci/NLMR/}
}
\usage{
classified_landscape
}
\description{
An example map to show landscapetools functionality
generated with the nlm_random() algorithm with factorial values.
}
\keyword{datasets}
