% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fbmmap}
\alias{fbmmap}
\title{Example map (fractional brownian motion).}
\format{A raster layer object.}
\source{
Simulated neutral landscape models with R. \url{https://github.com/ropensci/NLMR/}
}
\usage{
fbmmap
}
\description{
An example map to show landscapetools functionality
generated with the nlm_fbm() algorithm.
}
\keyword{datasets}
