% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gradient_landscape}
\alias{gradient_landscape}
\title{Example map (planar gradient).}
\format{A raster layer object.}
\source{
Simulated neutral landscape models with R. \url{https://github.com/ropensci/NLMR/}
}
\usage{
gradient_landscape
}
\description{
An example map to show landscapetools functionality
generated with the nlm_planargradient() algorithm.
}
\keyword{datasets}
