% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{random_landscape}
\alias{random_landscape}
\title{Example map (random).}
\format{A raster layer object.}
\source{
Simulated neutral landscape models with R. \url{https://github.com/ropensci/NLMR/}
}
\usage{
random_landscape
}
\description{
An example map to show landscapetools functionality
generated with the nlm_random() algorithm.
}
\keyword{datasets}
