% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rndmap}
\alias{rndmap}
\title{Example map (random).}
\format{A raster layer object.}
\source{
Simulated neutral landscape models with R. \url{https://github.com/marcosci/nlmr/}
}
\usage{
rndmap
}
\description{
An example map to show landscapetools functionality
generated with the nlm_random() algorithm.
}
\keyword{datasets}
