% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_as_integer.R
\name{util_as_integer}
\alias{util_as_integer}
\alias{util_as_integer.RasterLayer}
\title{util_as_integer}
\usage{
util_as_integer(x)

\method{util_as_integer}{RasterLayer}(x)
}
\arguments{
\item{x}{raster}
}
\value{
RasterLayer
}
\description{
Coerces raster values to integers
}
\details{
Coerces raster values to integers, which is sometimes needed if you want further
methods that rely on integer values.
}
\examples{
# Mode 1
util_as_integer(fractal_landscape)


}
