% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_facetplot.R
\name{util_facetplot}
\alias{util_facetplot}
\title{util_facetplot() for visual overview}
\usage{
util_facetplot(x, nrow = NULL, ncol = NULL)
}
\arguments{
\item{x}{Raster* Layer, Stack, Brick or a list of rasterLayers.}

\item{nrow, ncol}{Number of rows and columns.}
}
\value{
ggplot
}
\description{
Plot multiple raster (RasterStack, -brick or list of raster) side by side as facets.
}
\details{
The output uses ggplots faceting and beforehand raster2tibble.
Thus you will loose any spatial information (resolution, extent or reference system).
Only raw tiles are displayed and the number of cells determines the size of the plot.
This can lead to huge size differences between maps, but if you plot for example
multiple maps from a time series side by side it works as intended. Depending on the
size of the maps it is advisable to store the plot in an object and print it to
a file. This will help with compressing and rendering the image.
}
\examples{
\dontrun{
breaks <- c(0.3, 0.5, 0.7)
binary_maps <- util_binarize(fbmmap, breaks)
util_facetplot(binary_maps, nrow = 1)

lst <- list(fbmmap, grdmap, rndmap)
util_facetplot(lst)
}

}
