% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_merge.R
\name{util_merge}
\alias{util_merge}
\title{util_merge}
\usage{
util_merge(primary_nlm, secondary_nlm, scalingfactor = 1, rescale = TRUE)
}
\arguments{
\item{primary_nlm}{Primary \code{Raster* object}}

\item{secondary_nlm}{A list or stack of \code{Raster* object}s that are merged with the primary \code{Raster* object}}

\item{scalingfactor}{Weight for the secondary \code{Raster* objects}}

\item{rescale}{If \code{TRUE} (default), the values are rescaled between 0-1.}
}
\value{
Rectangular matrix with values ranging from 0-1
}
\description{
Merge a primary raster with other rasters weighted by scaling factors.
}
\examples{
(util_merge(grdmap, rndmap))

}
