% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscapetools.R
\docType{package}
\name{landscapetools-package}
\alias{landscapetools}
\alias{landscapetools-package}
\title{landscapetools}
\description{
\emph{landscapetools} is
}
\details{
...
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/marcosci/landscapetools}
  \item Report bugs at \url{https://github.com/marcosci/landscapetools/issues}
}

}
\author{
\strong{Maintainer}: Marco Sciaini \email{sciaini.marco@gmail.com}

Authors:
\itemize{
  \item Matthias Fritsch \email{matthias.fritsch@forst.uni-goettingen.de}
}

}
