% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/landscapetools.R
\docType{package}
\name{landscapetools-package}
\alias{landscapetools}
\alias{landscapetools-package}
\title{landscapetools}
\description{
\emph{landscapetools} provides utility functions to work with landscape data
(raster* Objects).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ropensci.github.io/landscapetools/}
  \item Report bugs at \url{https://github.com/ropensci/landscapetools/issues}
}

}
\author{
\strong{Maintainer}: Marco Sciaini \email{sciaini.marco@gmail.com} (0000-0002-3042-5435)

Authors:
\itemize{
  \item Matthias Fritsch \email{matthias.fritsch@forst.uni-goettingen.de}
  \item Maximillian H.K. Hesselbarth \email{maximilian.hesselbarth@uni-goettingen.de} (0000-0003-1125-9918)
  \item Jakub Nowosad \email{nowosad.jakub@gmail.com} (0000-0002-1057-3721)
}

Other contributors:
\itemize{
  \item Laura Graham (Laura reviewed the package for rOpenSci, see 
                  https://github.com/ropensci/onboarding/issues/188) [reviewer]
  \item Jeffrey Hollister (Jeffrey reviewed the package for rOpenSci, see 
                  https://github.com/ropensci/onboarding/issues/188) [reviewer]
}

}
