% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_landscape.R, R/util_plot.R
\name{show_landscape}
\alias{show_landscape}
\alias{show_landscape.RasterLayer}
\alias{show_landscape.list}
\alias{show_landscape.RasterStack}
\alias{show_landscape.RasterBrick}
\title{show_landscape}
\usage{
show_landscape(x, xlab, ylab, discrete, div_scales, n_col, n_row, ...)

\method{show_landscape}{RasterLayer}(x, xlab = "Easting",
  ylab = "Northing", discrete = FALSE, ...)

\method{show_landscape}{list}(x, xlab = "Easting", ylab = "Northing",
  discrete = FALSE, div_scales = FALSE, n_col = NULL, n_row = NULL,
  ...)

\method{show_landscape}{RasterStack}(x, xlab = "Easting",
  ylab = "Northing", discrete = FALSE, div_scales = FALSE,
  n_col = NULL, n_row = NULL, ...)

show_landscape(x, xlab, ylab, discrete, div_scales, n_col, n_row, ...)

\method{show_landscape}{RasterLayer}(x, xlab = "Easting",
  ylab = "Northing", discrete = FALSE, ...)

\method{show_landscape}{list}(x, xlab = "Easting", ylab = "Northing",
  discrete = FALSE, div_scales = FALSE, n_col = NULL, n_row = NULL,
  ...)

\method{show_landscape}{RasterStack}(x, xlab = "Easting",
  ylab = "Northing", discrete = FALSE, div_scales = FALSE,
  n_col = NULL, n_row = NULL, ...)

\method{show_landscape}{RasterBrick}(x, xlab = "Easting",
  ylab = "Northing", discrete = FALSE, div_scales = FALSE,
  n_col = NULL, n_row = NULL, ...)
}
\arguments{
\item{x}{Raster* object}

\item{xlab}{x axis label, default "Easting"}

\item{ylab}{y axis label, default "Northing"}

\item{discrete}{If TRUE, the function plots a raster with
a discrete legend.}

\item{div_scales}{multiple scales for facets}

\item{n_col}{number of cols if list, stack or brick ...}

\item{n_row}{number of rows if list, stack or brick ...}

\item{...}{Arguments for  \code{\link{theme_nlm}}}

\item{x}{Raster* object}

\item{xlab}{x axis label, default "Easting"}

\item{ylab}{y axis label, default "Northing"}

\item{discrete}{If TRUE, the function plots a raster with}

\item{...}{Arguments for  \code{\link{theme_nlm}}}
}
\value{
ggplot2 Object

ggplot2 Object
}
\description{
Plot a Raster* object with the NLMR default theme (as ggplot).

Plot a Raster* object with the NLMR default theme (as ggplot).
}
\examples{
\dontrun{
x <- grdmap

# classify
y <- util_classify(grdmap,  n = 3, level_names = c("Land Use 1", "Land Use 2", "Land Use 3"))

show_landscape(x)
show_landscape(y, discrete = TRUE)

show_landscape(list(grdmap, rndmap))
show_landscape(raster::stack(grdmap, rndmap))
}

\dontrun{
x <- grdmap

# classify
y <- util_classify(grdmap,  n = 3, level_names = c("Land Use 1", "Land Use 2", "Land Use 3"))

show_landscape(x)
show_landscape(y, discrete = TRUE)

show_landscape(list(grdmap, rndmap))
show_landscape(raster::stack(grdmap, rndmap))
}

}
